Auth state tests
