+++ test/unit/definitions/divorce/AuthorisationCaseState.test.js
@@ -1,11 +1,10 @@
 const { expect } = require('chai');
 const { uniqWith } = require('lodash');
-const { isFieldDuplicated, loadAllFiles } = require('../../utils/utils');
+const { isFieldDuplicated } = require('../../utils/utils');
 const { createAssertExists } = require('../../utils/assertBuilders');
+const { prod, nonprod } = require('../../utils/dataProvider');
 
 const assertStateExists = createAssertExists('State');
-const getAuthorisationCaseStateDefinitions = loadAllFiles('AuthorisationCaseState');
-const getStateDefinitions = loadAllFiles('State');
 
 describe('AuthorisationCaseState', () => {
   describe('NonProd files definitions:', () => {
@@ -13,26 +12,8 @@ describe('AuthorisationCaseState', () => {
     let nonProdStates = [];
 
     before(() => {
-      nonProd = getAuthorisationCaseStateDefinitions(
-        [
-          'AuthorisationCaseState',
-          'AuthorisationCaseState-alternative-service-nonprod',
-          'AuthorisationCaseState-alt-service-process-server-nonprod',
-          'AuthorisationCaseState-bailiff-nonprod',
-          'AuthorisationCaseState-deemed-and-dispensed-nonprod',
-          'AuthorisationCaseState-general-referral-nonprod',
-          'AuthorisationCaseState-share-a-case-nonprod',
-          'AuthorisationCaseState-nonprod'
-        ]
-      );
-      nonProdStates = getStateDefinitions([
-        'State',
-        'State-alternative-service-nonprod',
-        'State-alt-service-process-server-nonprod',
-        'State-bailiff-nonprod',
-        'State-deemed-and-dispensed-nonprod',
-        'State-general-referral-nonprod'
-      ]);
+      nonProd = nonprod.AuthorisationCaseState;
+      nonProdStates = nonprod.State;
     });
 
     it('should contain a unique case state, case type ID and role (no duplicates) for nonprod files', () => {
@@ -46,26 +27,21 @@ describe('AuthorisationCaseState', () => {
   });
 
   describe('Prod files definitions:', () => {
-    let prod = [];
+    let prodOnly = [];
     let prodStates = [];
 
     before(() => {
-      prod = getAuthorisationCaseStateDefinitions(
-        [
-          'AuthorisationCaseState',
-          'AuthorisationCaseState-prod'
-        ]
-      );
-      prodStates = getStateDefinitions(['State']);
+      prodOnly = prod.AuthorisationCaseState;
+      prodStates = prod.State;
     });
 
     it('should contain a unique case state, case type ID and role (no duplicates)', () => {
-      const uniqResult = uniqWith(prod, isFieldDuplicated('CaseStateID'));
-      expect(uniqResult).to.eql(prod);
+      const uniqResult = uniqWith(prodOnly, isFieldDuplicated('CaseStateID'));
+      expect(uniqResult).to.eql(prodOnly);
     });
 
     it('should use existing states ', () => {
-      assertStateExists(prod, prodStates);
+      assertStateExists(prodOnly, prodStates);
     });
   });
 });
